requirements.txt
Cython==0.29.23
Flask==1.1.2
Flask-Cors==3.0.10
Jinja2==3.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.3.4
MarkupSafe==2.0.1
Pillow==8.2.0
PyJWT==1.7.1
PySocks==1.7.1
Werkzeug==1.0.1
absl-py==0.12.0
aiohttp==3.7.4
astunparse==1.6.3
async-timeout==	3.0.1
attrs==	21.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==8.0.1
coverage==5.5
cryptography==3.4.7
gast==0.4.0
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.36.1
h5py==2.10.0
idna==2.10
importlib-metadata==3.10.0
itsdangerous==2.0.1
mkl-fft==1.3.0
mkl-random==1.2.1
mkl-service==2.3.0
multidict==5.1.0
numpy==1.20.2
oauthlib==3.1.0
olefile==0.46
opt-einsum==3.3.0
pip==21.1.1
protobuf==3.14.0
pyOpenSSL==20.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyreadline==2.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.6.2
setuptools==52.0.0.post20210125
six	==1.15.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.6.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.4
wheel==0.36.2
win-inet-pton==1.1.0
wincertstore==0.2